services/random-name/requirements.txt
python_random_name_generator==0.0.6
